Move sending to a dedicatd thread.
